pyproject.toml
[project]
name = "coxwave-openai-hackerton"
version = "0.1.0"
description = "Video generation AI agent system for marketing videos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.6.9",
]

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.400",
    "pre-commit>=4.0.0",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
